backend/requirements.txt
fastapi
uvicorn[standard]
openai
chromadb
sentence-transformers
python-dotenv
pydantic
requests
aiofiles
python-multipart
